fsx: Move some functions around.
